fix(remote-access): address code-review findings
- cloud/tenants.rs: in mint_token, mark the box entitled *before* parsing the
  token body. A non-403 success means the subscription is active, so a malformed
  (e.g. truncated) 200 must still clear a suspended flag — otherwise the box
  could never self-heal, since every re-probe would hit the same parse error.
- Logging: include the structured field values in the message text per
  .agents/logging.md (ddns/mod.rs tenant_id, ddns/runner.rs error,
  ddns/region.rs error), matching the tls/runner.rs precedent.
- admin-site: extract the shared wardnet/Cloudflare enrollment state machine
  into useWardnetEnrollment (lib/wardnet-enrollment.ts) plus shared field
  components (components/features/wardnet-enrollment.tsx). RemoteAccess and
  Step7RemoteAccess previously duplicated the state, debounced availability
  check, handlers, and form/hint components verbatim. Each step button now gates
  only on its own mutation's pending flag instead of a single global `busy`, so
  an unrelated in-flight request no longer disables the current step.
